+++ src/Main.svelte
@@ -46,23 +46,50 @@
     <div class="w-full h-full min-h-screen bg-white dark:bg-gray-800">
       {#each $feedData.features as feature (feature.properties.code)}
         <div
-          class="flex items-center justify-between border-b dark:border-gray-600 hover:bg-gray-100 dark:hover:bg-gray-900 p-2"
+          class="flex items-center justify-between border-b dark:border-gray-600 hover:bg-gray-100 dark:hover:bg-gray-900 py-1"
         >
-          <div class="w-full flex items-center gap-4">
-            <div class="w-8 text-center">
-              <span class:font-bold={feature.properties.mag >= 4.5} class:text-red-600={feature.properties.mag >= 7}>
-                {round1(feature.properties.mag)}
-              </span>
-            </div>
-            <div class="flex-1">
-              <p class:font-bold={feature.properties.mag >= 4.5} class:text-red-600={feature.properties.mag >= 7}>
-                {feature.properties.place}
-              </p>
-              <small class="text-xs text-gray-600 dark:text-gray-400">{timestampToLocalString(feature.properties.time)}</small>
-            </div>
+          <!-- Magnitude -->
+          <div class="px-4">
+            <span class:font-bold={feature.properties.mag >= 4.5} class:text-red-600={feature.properties.mag >= 7}>
+              {round1(feature.properties.mag)}
+            </span>
           </div>
 
-          <div class="w-24 text-right">
+          <!-- Location -->
+          <div class="flex-1 w-full flex flex-col">
+            <p class="flex items-center gap-2">
+              <span class:font-bold={feature.properties.mag >= 4.5} class:text-red-600={feature.properties.mag >= 7}
+                >{feature.properties.place}</span
+              >
+              <a
+                href={`http://geojson.io/#data=data:application/json,${encodeURIComponent(JSON.stringify(feature))}`}
+                target="_blank"
+                rel="noreferrer"
+                class="text-blue-600"
+                title="View on geojson.io"
+                use:tooltip={{ theme: "dark-border" }}
+              >
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  width="15"
+                  height="15"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-width="2"
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                  ><path d="M21 10c0 7-9 13-9 13s-9-6-9-13a9 9 0 0 1 18 0z" /><circle cx="12" cy="10" r="3" /></svg
+                >
+              </a>
+            </p>
+            <small class="text-xs text-gray-600 dark:text-gray-400">
+              {timestampToLocalString(feature.properties.time)}
+            </small>
+          </div>
+
+          <!-- Depth -->
+          <div class="w-24 px-2 text-right">
             <small class="text-xs text-gray-600 dark:text-gray-400" title="Depth">
               {round1(feature.geometry.coordinates[2])} km
             </small>